move 'contains' tests to separate file
